Add stop area information to stop page
